+++ src/splayTable.c
@@ -1096,7 +1096,7 @@ void inputSequenceArrayIntoSplayTableAndArchive(ReadSet * reads,
 
 	gettimeofday(&end, NULL);
 	timersub(&end, &start, &diff);
-	velvetLog(" === Sequences loaded in %ld.%06ld s\n", diff.tv_sec, diff.tv_usec);
+	velvetLog(" === Sequences loaded in %ld.%06ld s\n", (long) diff.tv_sec, (long) diff.tv_usec);
 
 	fclose(outfile);
